refactor(app): split home section views
